unify comments style and remove debug codes
