+++ nni/retiarii/oneshot/pytorch/__init__.py
@@ -5,7 +5,6 @@ from .darts import DartsTrainer
 from .enas import EnasTrainer
 from .proxyless import ProxylessTrainer
 from .random import SinglePathTrainer, RandomTrainer
-from .utils import replace_input_choice, replace_layer_choice
-from .differentiable import DartsModel, ProxylessModel
-from .sampling import EnasModel,RandomSampleModel
-from .base_lightning import MergeTrainValDataset
+from .differentiable import DartsModule, ProxylessModule
+from .sampling import EnasModule, RandomSampleModule
+from .utils import ParallelTrainValDataLoader, ConcatenateTrainValDataLoader
